QA'd through otel in action
