+++ docs/apps/opentelemetry/weather_app/app-demo-all-suppressed.py
@@ -1,14 +1,13 @@
 # Setup:
 # 1. pip install shiny chatlas logfire requests boto3 opentelemetry-instrumentation-anthropic
-# 2. Configure your AI provider of choice (here we use AWS Bedrock but you can replace
-#    with the provider and auth setup of your choice)
+# 2. Configure your AI provider of choice (here we use AWS Bedrock)
 # 3. Set environment variables:
 #    export SHINY_OTEL_COLLECT=reactivity
 #    export AWS_PROFILE=claude (For AWS Bedrock, you may have an API key instead)
 # 4. Optional: Configure Logfire for observability:
 #    - Run: logfire configure
 #    - View traces at https://logfire.pydantic.dev/
-# 5. Run: shiny run app-demo-all-express.py
+# 5. Run: shiny run app-demo-all-core.py
 
 import os
 
@@ -25,13 +24,14 @@ from chatlas import ChatBedrockAnthropic
 
 # Instrumentation for our chosen chat client (Anthropic/Claude via chatlas)
 from opentelemetry.instrumentation.anthropic import AnthropicInstrumentor
-from shiny.express import ui
 
 # Configure Logfire (optional)
-_logfire = logfire.configure()
+logfire.configure()
 
 # Instrument chatlas/Anthropic for automatic tracing
 AnthropicInstrumentor().instrument()
+
+from shiny import App, otel, ui
 
 
 def get_weather_forecast(lat: float, lon: float) -> str:
@@ -75,28 +75,39 @@ def get_weather_forecast(lat: float, lon: float) -> str:
         return error_msg
 
 
-# Initialize Claude client
-chat_client = ChatBedrockAnthropic(
-    system_prompt="Be terse.", model="us.anthropic.claude-sonnet-4-5-20250929-v1:0"
+app_ui = ui.page_fillable(
+    ui.chat_ui(
+        id="chat",
+        messages=["Ask me about the weather! Try: What is the weather in Atlanta, GA?"],
+    ),
+    fillable_mobile=True,
 )
-# Register the weather tool we defined above
-chat_client.register_tool(get_weather_forecast)
-
-# Set page options
-ui.page_opts(fillable=True, fillable_mobile=True)
-
-# Create a chat instance
-chat = ui.Chat(id="chat")
-chat.ui(messages=["Ask me about the weather! Try: What is the weather in Atlanta, GA?"])
 
 
-@chat.on_user_submit
-async def handle_user_input(user_input: str):
-    """
-    Handle user message submission.
-    Errors are automatically displayed as notifications by Shiny.
-    """
-    # Send message to Claude and stream response
-    # AnthropicInstrumentor automatically traces this call for Otel
-    response = await chat_client.stream_async(user_input)
-    await chat.append_message_stream(response)
+def server(input, output, session):
+    # Initialize Claude client
+    chat_client = ChatBedrockAnthropic(
+        system_prompt="Be terse.", model="us.anthropic.claude-sonnet-4-5-20250929-v1:0"
+    )
+    # Register the weather tool we defined above
+    chat_client.register_tool(get_weather_forecast)
+
+    # TODO: Before after compare this
+    # Create a chat instance
+    with otel.suppress():
+        chat = ui.Chat(id="chat")
+
+    @chat.on_user_submit
+    async def handle_user_input(user_input: str):
+        """
+        Handle user message submission.
+        Errors are automatically displayed as notifications by Shiny.
+        """
+        # Send message to Claude and stream response
+        # AnthropicInstrumentor automatically traces this call for Otel
+        with otel.suppress():
+            response = await chat_client.stream_async(user_input)
+            await chat.append_message_stream(response)
+
+
+app = App(app_ui, server)
